requirements.txt
python-pptx
Pillow
requests
tqdm
transformers
sentencepiece
PyMuPDF
gTTS
accelerate
pytest
pytest-cov
# optional (dev)
black
ruff
